update test_parallel_processing to get all specs from config
