+++ tests/test_config.ini
@@ -76,4 +76,27 @@ ncoefs   = 10
 
 
 
+[LOAD_MODULES]
+
+relative__processfuncs = parallel_processing_config.py
+
+
+
+[PROCESS_SPECS]
+
+finalout_name = results.h5
+dumpfolder = dump01
+
+path_relative__save_path = proc_test
+
+
+# add some additional values
+float__f0 = 1245
+float__f1 = 5.4
+int__i0 = 1
+int__i1 = 5
+bool_b0 = False
+bool_b1 = True
+
+
